NMS-13090: Build prototype of configuration system for vacuumd config: move into its own module
