+++ opennms-config-tester/src/main/resources/META-INF/opennms/applicationContext-configTester.xml
@@ -507,6 +507,7 @@
 
   <bean id="userFactory" class="org.opennms.netmgt.config.UserFactory" depends-on="userFactory-init" factory-method="getInstance"/>
 
+  <!-- TODO: Patrick -->
   <bean id="vacuumdConfigFactory-init" class="org.springframework.beans.factory.config.MethodInvokingFactoryBean">
     <property name="staticMethod"><value>org.opennms.netmgt.config.VacuumdConfigWrapper.init</value></property>
   </bean>
