+++ opennms-services/src/main/java/org/opennms/netmgt/vacuumd/Vacuumd.java
@@ -43,13 +43,13 @@ import org.opennms.core.db.DataSourceFactory;
 import org.opennms.core.logging.Logging;
 import org.opennms.core.utils.ConfigFileConstants;
 import org.opennms.netmgt.config.VacuumdConfigWrapper;
-import org.opennms.netmgt.config.service.ConfigurationNotAvailableException;
-import org.opennms.netmgt.config.service.ConfigurationService;
 import org.opennms.netmgt.config.vacuumd.Action;
 import org.opennms.netmgt.config.vacuumd.Automation;
 import org.opennms.netmgt.config.vacuumd.Statement;
 import org.opennms.netmgt.config.vacuumd.Trigger;
 import org.opennms.netmgt.config.vacuumd.VacuumdConfiguration;
+import org.opennms.netmgt.configservice.ConfigurationNotAvailableException;
+import org.opennms.netmgt.configservice.ConfigurationService;
 import org.opennms.netmgt.daemon.AbstractServiceDaemon;
 import org.opennms.netmgt.events.api.EventConstants;
 import org.opennms.netmgt.events.api.EventIpcManager;
